DON'T MERGE - This just captures the hackaround effort required to achieve JSON roundtrips of plugin data
The PluginTester was already able to serialize a FieldVisitInfo DTO to JSON. It's been doing that fine for over a year.

This commit represents all workaround hackery required to allow that JSON to be deserialized from JSON back into object form again.

Round-tripping the FieldVisitInfo DTO to/from JSON is a critical feature lacking in the framework.